Add memmove and nmemmove
Note: I decided against adding memcpy, since the manpage pointed out
that many bugs have been caused by not ensuring that the source and
destination is non-overlapping.
I don't think the performance loss is significant (and if you *need*
that type of performance, why the hell are you using this rando's hobby
project in the first place?), and anyways you *should* be able to just
do a quick check at the start of the function for that and apply
optimizations conditionally, no? Or maybe I'm wrong...

Also, I just realised I didn't `restrict` the parameters to memmove,
will that be a problem? I assume not, but I'll google quick and make a
new commit if it is (idk how restrict works)