Fix bugs in fscan* functions
There were two bugs in the fscan* functions, first some parentheses
around a set-and-compare sequence was misplaced, resulting in the
space-skipping loop not working correctly (skipping *all* characters
until an EOF is encountered). Secondly, I implemented scanning for
signed and unsigned integers the same, not allowing for the scanning of
negative values.